apps/site-vitrine/backend/requirements.txt
fastapi
email-validator
uvicorn[standard]
python-dotenv
anthropic
httpx

# DATABASE
sqlalchemy
psycopg2-binary

passlib[bcrypt]
python-jose[cryptography]
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart
slowapi
